# core: Add migrating deframer