--- a/core/src/main/java/io/grpc/internal/AbstractStream.java
+++ b/core/src/main/java/io/grpc/internal/AbstractStream.java
@@ -41,8 +41,18 @@ public abstract class AbstractStream implements Stream {
   protected abstract TransportState transportState();
 
   @Override
+  public void optimizeForDirectExecutor() {
+    transportState().optimizeForDirectExecutor();
+  }
+
+  @Override
   public final void setMessageCompression(boolean enable) {
     framer().setMessageCompression(enable);
+  }
+
+  @Override
+  public final void request(int numMessages) {
+    transportState().requestMessagesFromDeframer(numMessages);
   }
 
   @Override
@@ -112,6 +122,7 @@ public abstract class AbstractStream implements Stream {
     private final Object onReadyLock = new Object();
     private final StatsTraceContext statsTraceCtx;
     private final TransportTracer transportTracer;
+    private final MessageDeframer rawDeframer;
 
     /**
      * The number of bytes currently queued, waiting to be sent. When this falls below
@@ -138,17 +149,23 @@ public abstract class AbstractStream implements Stream {
         TransportTracer transportTracer) {
       this.statsTraceCtx = checkNotNull(statsTraceCtx, "statsTraceCtx");
       this.transportTracer = checkNotNull(transportTracer, "transportTracer");
-      deframer = new MessageDeframer(
+      rawDeframer = new MessageDeframer(
           this,
           Codec.Identity.NONE,
           maxMessageSize,
           statsTraceCtx,
           transportTracer);
+      deframer = new MigratingThreadDeframer(this, this, rawDeframer);
+    }
+
+    final void optimizeForDirectExecutor() {
+      rawDeframer.setListener(this);
+      deframer = rawDeframer;
     }
 
     protected void setFullStreamDecompressor(GzipInflatingBuffer fullStreamDecompressor) {
-      deframer.setFullStreamDecompressor(fullStreamDecompressor);
-      deframer = new ApplicationThreadDeframer(this, this, (MessageDeframer) deframer);
+      rawDeframer.setFullStreamDecompressor(fullStreamDecompressor);
+      deframer = new ApplicationThreadDeframer(this, this, rawDeframer);
     }
 
     final void setMaxInboundMessageSize(int maxSize) {
@@ -197,15 +214,34 @@ public abstract class AbstractStream implements Stream {
     }
 
     /**
-     * Called to request the given number of messages from the deframer. Must be called from the
-     * transport thread.
+     * Called to request the given number of messages from the deframer. May be called from any
+     * thread.
      */
-    public final void requestMessagesFromDeframer(final int numMessages) {
-      try {
+    private void requestMessagesFromDeframer(final int numMessages) {
+      if (deframer instanceof ThreadOptimizedDeframer) {
         deframer.request(numMessages);
-      } catch (Throwable t) {
-        deframeFailed(t);
+        return;
       }
+      class RequestRunnable implements Runnable {
+        @Override public void run() {
+          try {
+            deframer.request(numMessages);
+          } catch (Throwable t) {
+            deframeFailed(t);
+          }
+        }
+      }
+
+      runOnTransportThread(new RequestRunnable());
+    }
+
+    /**
+     * Very rarely used. Prefer stream.request() instead of this; this method is only necessary if
+     * a stream is not available.
+     */
+    @VisibleForTesting
+    public final void requestMessagesFromDeframerForTesting(int numMessages) {
+      requestMessagesFromDeframer(numMessages);
     }
 
     public final StatsTraceContext getStatsTraceContext() {