--- a/core/src/main/java/io/grpc/internal/NoopClientStream.java
+++ b/core/src/main/java/io/grpc/internal/NoopClientStream.java
@@ -67,6 +67,9 @@ public class NoopClientStream implements ClientStream {
   }
 
   @Override
+  public void optimizeForDirectExecutor() {}
+
+  @Override
   public void setCompressor(Compressor compressor) {}
 
   @Override