… core/src/main/java/io/grpc/internal/MigratingThreadDeframer.java
… core/src/test/java/io/grpc/internal/AbstractClientStreamTest.java
… core/src/test/java/io/grpc/internal/ApplicationThreadDeframerTest.java
… netty/src/main/java/io/grpc/netty/NettyClientStream.java
… netty/src/main/java/io/grpc/netty/NettyServerStream.java
… stub/src/main/java/io/grpc/stub/ClientCalls.java