Possible alternative to #5800 which also helps reading many small messages. Not a direct replacement because onHalfClose() still delays the server from responding earlier when using our stubs.

CC @carl-mastrangelo 

I hacked the server stub to avoid waiting for onHalfClose for these numbers:
```
Before
After
Benchmark                                               (direct)  (transport)    Mode    Cnt        Score     Error  Units
Benchmark                                               (direct)  (transport)    Mode    Cnt         Score     Error  Units
TransportBenchmark.unaryCall1024                           false        NETTY  sample  55343   180432.321 ± 960.495  ns/op
TransportBenchmark.unaryCall1024                           false        NETTY  sample  63613   156958.498 ± 996.485  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false        NETTY  sample          110464.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false        NETTY  sample           97536.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false        NETTY  sample          173568.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false        NETTY  sample          150528.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false        NETTY  sample          220672.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false        NETTY  sample          194816.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false        NETTY  sample          240384.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false        NETTY  sample          214528.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false        NETTY  sample          289280.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false        NETTY  sample          263168.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false        NETTY  sample          401887.232            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false        NETTY  sample          357259.264            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false        NETTY  sample         1002604.134            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false        NETTY  sample          946923.930            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false        NETTY  sample         9093120.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false        NETTY  sample        10747904.000            ns/op
```