@carl-mastrangelo 2019-06-02
I ran the tracer with your change and attached the results. 

[post.zip](https://github.com/grpc/grpc-java/files/3244480/post.zip)
[pre.zip](https://github.com/grpc/grpc-java/files/3244481/pre.zip)