@ejona86 2020-06-17
With the just-pushed changes, I think I have fixed potential issues and resolved virtually all performance issues. We see very large gains for Netty (2x for blocking, 3x for async). The only benchmark that decreased is Netty+direct+unary latency which is ~3.5µs slower:

```
TransportBenchmark.unaryCall1024Latency                                   true        NETTY  sample   79273      125957.104 ±       688.071  ns/op
TransportBenchmark.unaryCall1024Latency                                   true        NETTY  sample   77243      129275.584 ±       793.104  ns/op
```

If I unconditionally set `callExecutorIsDirect = true` in `ClientCallImpl`, then the performance decrease disappears (because this benchmark uses blocking, so direct isn't used on client-side). I think we should just accept this decrease, given the huge gains provided elsewhere and the rarity of direct executor.

```
Before
After
Benchmark                                                             (direct)  (transport)    Mode     Cnt           Score           Error  Units
Benchmark                                                             (direct)  (transport)    Mode     Cnt           Score           Error  Units
TransportBenchmark.streamingCallsByteThroughput                           true    INPROCESS   thrpt      10  9987578604.232 ±  24202529.991  ops/s
TransportBenchmark.streamingCallsByteThroughput                           true    INPROCESS   thrpt      10  9775020244.225 ±  27557060.688  ops/s
TransportBenchmark.streamingCallsByteThroughput                           true        NETTY   thrpt      10   877382865.222 ±  29590657.813  ops/s
TransportBenchmark.streamingCallsByteThroughput                           true        NETTY   thrpt      10   859303134.930 ±  22487705.398  ops/s
TransportBenchmark.streamingCallsByteThroughput                           true       OKHTTP   thrpt      10   410195297.480 ±   7139911.699  ops/s
TransportBenchmark.streamingCallsByteThroughput                           true       OKHTTP   thrpt      10   408359041.153 ±  18808969.171  ops/s
TransportBenchmark.streamingCallsByteThroughput                          false    INPROCESS   thrpt      10  9826245515.349 ± 139614657.375  ops/s
TransportBenchmark.streamingCallsByteThroughput                          false    INPROCESS   thrpt      10  9845838238.437 ±  58984357.716  ops/s
TransportBenchmark.streamingCallsByteThroughput                          false        NETTY   thrpt      10  1374727164.633 ±  61962267.634  ops/s
TransportBenchmark.streamingCallsByteThroughput                          false        NETTY   thrpt      10  1365981174.137 ±  45913840.058  ops/s
TransportBenchmark.streamingCallsByteThroughput                          false       OKHTTP   thrpt      10   534065331.852 ±  12816316.465  ops/s
TransportBenchmark.streamingCallsByteThroughput                          false       OKHTTP   thrpt      10   531680014.121 ±  21812113.309  ops/s
TransportBenchmark.streamingCallsMessageThroughput                        true    INPROCESS   thrpt      10    13095457.490 ±    290143.597  ops/s
TransportBenchmark.streamingCallsMessageThroughput                        true    INPROCESS   thrpt      10    16891454.791 ±    301832.219  ops/s
TransportBenchmark.streamingCallsMessageThroughput                        true        NETTY   thrpt      10      338443.051 ±     14308.936  ops/s
TransportBenchmark.streamingCallsMessageThroughput                        true        NETTY   thrpt      10      698802.019 ±     16166.566  ops/s
TransportBenchmark.streamingCallsMessageThroughput                        true       OKHTTP   thrpt      10      844963.002 ±     39991.773  ops/s
TransportBenchmark.streamingCallsMessageThroughput                        true       OKHTTP   thrpt      10      964963.963 ±     34750.486  ops/s
TransportBenchmark.streamingCallsMessageThroughput                       false    INPROCESS   thrpt      10      619848.174 ±      6627.489  ops/s
TransportBenchmark.streamingCallsMessageThroughput                       false    INPROCESS   thrpt      10     1218679.299 ±     44057.306  ops/s
TransportBenchmark.streamingCallsMessageThroughput                       false        NETTY   thrpt      10      344388.670 ±     20630.013  ops/s
TransportBenchmark.streamingCallsMessageThroughput                       false        NETTY   thrpt      10      704464.722 ±     24627.621  ops/s
TransportBenchmark.streamingCallsMessageThroughput                       false       OKHTTP   thrpt      10      933180.566 ±     25001.009  ops/s
TransportBenchmark.streamingCallsMessageThroughput                       false       OKHTTP   thrpt      10     1298180.172 ±     30323.983  ops/s
TransportBenchmark.streamingAsyncCallsMessageThroughput                  false    INPROCESS   thrpt      10      666894.487 ±      9662.139  ops/s
TransportBenchmark.streamingAsyncCallsMessageThroughput                  false    INPROCESS   thrpt      10      670339.596 ±     26359.860  ops/s
TransportBenchmark.streamingAsyncCallsMessageThroughput                  false        NETTY   thrpt      10      334472.985 ±     31568.403  ops/s
TransportBenchmark.streamingAsyncCallsMessageThroughput                  false        NETTY   thrpt      10      901221.922 ±     46658.088  ops/s
TransportBenchmark.streamingAsyncCallsMessageThroughput                  false       OKHTTP   thrpt      10      802237.671 ±      9310.184  ops/s
TransportBenchmark.streamingAsyncCallsMessageThroughput                  false       OKHTTP   thrpt      10     1123479.496 ±     37560.982  ops/s
TransportBenchmark.unaryCallsByteThroughput                               true    INPROCESS   thrpt      10  9911178501.741 ±  25436686.349  ops/s
TransportBenchmark.unaryCallsByteThroughput                               true    INPROCESS   thrpt      10  9576898669.805 ±  11741708.546  ops/s
TransportBenchmark.unaryCallsByteThroughput                               true        NETTY   thrpt      10   845501206.005 ±  23452363.207  ops/s
TransportBenchmark.unaryCallsByteThroughput                               true        NETTY   thrpt      10   882896731.619 ±  48600892.522  ops/s
TransportBenchmark.unaryCallsByteThroughput                               true       OKHTTP   thrpt      10   395183239.133 ±  11183757.763  ops/s
TransportBenchmark.unaryCallsByteThroughput                               true       OKHTTP   thrpt      10   402076474.053 ±   7282599.596  ops/s
TransportBenchmark.unaryCallsByteThroughput                              false    INPROCESS   thrpt      10  9851110309.288 ±  62265567.470  ops/s
TransportBenchmark.unaryCallsByteThroughput                              false    INPROCESS   thrpt      10  9758015980.918 ±  73689391.675  ops/s
TransportBenchmark.unaryCallsByteThroughput                              false        NETTY   thrpt      10  1294748876.721 ±  72386130.209  ops/s
TransportBenchmark.unaryCallsByteThroughput                              false        NETTY   thrpt      10  1275806216.690 ±  49310207.007  ops/s
TransportBenchmark.unaryCallsByteThroughput                              false       OKHTTP   thrpt      10   526216357.867 ±  29111355.493  ops/s
TransportBenchmark.unaryCallsByteThroughput                              false       OKHTTP   thrpt      10   512821807.915 ±  11923335.566  ops/s
TransportBenchmark.unaryCall1024Latency                                   true    INPROCESS  sample  243001        2695.254 ±        67.836  ns/op
TransportBenchmark.unaryCall1024Latency                                   true    INPROCESS  sample  246090        2633.125 ±        54.907  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00        true    INPROCESS  sample                2188.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00        true    INPROCESS  sample                2184.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50        true    INPROCESS  sample                2496.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50        true    INPROCESS  sample                2456.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90        true    INPROCESS  sample                2744.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90        true    INPROCESS  sample                2676.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95        true    INPROCESS  sample                2840.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95        true    INPROCESS  sample                2780.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99        true    INPROCESS  sample                4664.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99        true    INPROCESS  sample                4536.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999       true    INPROCESS  sample               19040.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999       true    INPROCESS  sample               18781.088                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999      true    INPROCESS  sample               31897.510                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999      true    INPROCESS  sample               30329.018                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00        true    INPROCESS  sample             1554432.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00        true    INPROCESS  sample             1939456.000                  ns/op
TransportBenchmark.unaryCall1024Latency                                   true        NETTY  sample   79273      125957.104 ±       688.071  ns/op
TransportBenchmark.unaryCall1024Latency                                   true        NETTY  sample   77243      129275.584 ±       793.104  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00        true        NETTY  sample               78208.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00        true        NETTY  sample               81792.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50        true        NETTY  sample              120192.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50        true        NETTY  sample              122624.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90        true        NETTY  sample              151552.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90        true        NETTY  sample              156160.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95        true        NETTY  sample              165632.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95        true        NETTY  sample              168448.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99        true        NETTY  sample              201538.560                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99        true        NETTY  sample              204800.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999       true        NETTY  sample              278016.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999       true        NETTY  sample              279040.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999      true        NETTY  sample              385300.992                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999      true        NETTY  sample              365709.107                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00        true        NETTY  sample             9469952.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00        true        NETTY  sample            10534912.000                  ns/op
TransportBenchmark.unaryCall1024Latency                                   true       OKHTTP  sample   76275      130909.083 ±      1031.559  ns/op
TransportBenchmark.unaryCall1024Latency                                   true       OKHTTP  sample   76293      130877.714 ±      1077.560  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00        true       OKHTTP  sample               76928.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00        true       OKHTTP  sample               70784.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50        true       OKHTTP  sample              123392.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50        true       OKHTTP  sample              124160.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90        true       OKHTTP  sample              157696.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90        true       OKHTTP  sample              157952.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95        true       OKHTTP  sample              172288.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95        true       OKHTTP  sample              172288.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99        true       OKHTTP  sample              230400.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99        true       OKHTTP  sample              233216.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999       true       OKHTTP  sample              375189.504                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999       true       OKHTTP  sample              391529.472                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999      true       OKHTTP  sample              752498.688                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999      true       OKHTTP  sample              800751.002                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00        true       OKHTTP  sample            11288576.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00        true       OKHTTP  sample            12959744.000                  ns/op
TransportBenchmark.unaryCall1024Latency                                  false    INPROCESS  sample  208415       16799.769 ±        64.351  ns/op
TransportBenchmark.unaryCall1024Latency                                  false    INPROCESS  sample  176386       14218.949 ±       103.459  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00       false    INPROCESS  sample                9328.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00       false    INPROCESS  sample                8056.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50       false    INPROCESS  sample               15008.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50       false    INPROCESS  sample               12752.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90       false    INPROCESS  sample               22464.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90       false    INPROCESS  sample               17760.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95       false    INPROCESS  sample               24096.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95       false    INPROCESS  sample               22080.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99       false    INPROCESS  sample               31456.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99       false    INPROCESS  sample               29824.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999      false    INPROCESS  sample               70752.256                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.999      false    INPROCESS  sample               69276.928                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999     false    INPROCESS  sample              132278.272                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.9999     false    INPROCESS  sample              154987.392                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00       false    INPROCESS  sample             2142208.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p1.00       false    INPROCESS  sample             3915776.000                  ns/op
TransportBenchmark.unaryCall1024Latency                                  false        NETTY  sample   62382      160067.441 ±       789.991  ns/op
TransportBenchmark.unaryCall1024Latency                                  false        NETTY  sample   63563      157087.757 ±      1086.319  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00       false        NETTY  sample               98560.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.00       false        NETTY  sample              102784.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50       false        NETTY  sample              154112.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.50       false        NETTY  sample              149760.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90       false        NETTY  sample              190464.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.90       false        NETTY  sample              191488.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95       false        NETTY  sample              205568.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.95       false        NETTY  sample              207872.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99       false        NETTY  sample              248064.000                  ns/op
TransportBenchmark.unaryCall1024Latency:unaryCall1024Latency·p0.99       false        NETTY  sample              251648.000                  ns/op
…